lt-16 / fix forum folder structure
